refactor(android): move most js ti extensions to common
