+++ common/Resources/ti.internal/extensions/ti/index.js
@@ -1,3 +1,7 @@
 // Load extensions to polyfill our own APIs
+import './ti.activity';
+import './ti.app.properties';
 import './ti.blob';
-import './ti.ui';
+import './ti.locale';
+import './ti.network.httpclient';
+import './ui';
