+++ common/Resources/ti.internal/extensions/ti/ti.network.httpclient.js
@@ -4,28 +4,24 @@
  * Licensed under the terms of the Apache Public License
  * Please see the LICENSE included with this distribution for details.
  */
-'use strict';
+/* globals OS_ANDROID */
+import PersistentHandle from './persistentHandle';
 
-var PersistentHandle = require('network').PersistentHandle;
+if (OS_ANDROID) {
+	const HTTPClient = Titanium.Network.HTTPClient;
 
-var TAG = 'HTTPClient';
-
-exports.bootstrap = function (Titanium) {
-
-	var HTTPClient = Titanium.Network.HTTPClient;
-
-	var _send = HTTPClient.prototype.send;
+	const _send = HTTPClient.prototype.send;
 	HTTPClient.prototype.send = function (options) {
 		// Retain the httpclient until the request has been finished.
-		var handle = new PersistentHandle(this);
+		const handle = new PersistentHandle(this);
 
 		this.on('disposehandle', function () {
 			handle.dispose();
 			if (kroll.DBG) {
-				kroll.log(TAG, 'The persistent handle is disposed.');
+				kroll.log('HTTPClient', 'The persistent handle is disposed.');
 			}
 		});
 
 		_send.call(this, options);
 	};
-};
+}
