+++ common/Resources/ti.internal/extensions/ti/ui/index.js
@@ -1,0 +1,9 @@
+import './iphone';
+import './listview';
+import './navigationwindow';
+import './semanticColor';
+import './tab';
+import './tabgroup';
+import './view';
+import './webview';
+import './window';
